Backend 🔚
- implemented controller
- converter added

+++ backend/UnitConverter/server/src/main/kotlin/com/pradyotprakash/unitconverter/features/converter/resource/ConverterResource.kt
@@ -1,12 +1,12 @@
 package com.pradyotprakash.unitconverter.features.converter.resource
 
-import com.pradyotprakash.unitconverter.Constants.Paths.Converter.CONVERT
-import com.pradyotprakash.unitconverter.Constants.Paths.Converter.LENGTH
-import com.pradyotprakash.unitconverter.Constants.Paths.Converter.WEIGHT
-import com.pradyotprakash.unitconverter.Constants.Paths.Converter.TEMPERATURE
-import com.pradyotprakash.unitconverter.LengthTypes
-import com.pradyotprakash.unitconverter.TemperatureTypes
-import com.pradyotprakash.unitconverter.WeightTypes
+import com.pradyotprakash.unitconverter.utils.Constants.Paths.Converter.CONVERT
+import com.pradyotprakash.unitconverter.utils.Constants.Paths.Converter.LENGTH
+import com.pradyotprakash.unitconverter.utils.Constants.Paths.Converter.WEIGHT
+import com.pradyotprakash.unitconverter.utils.Constants.Paths.Converter.TEMPERATURE
+import com.pradyotprakash.unitconverter.core.models.request.LengthTypes
+import com.pradyotprakash.unitconverter.core.models.request.TemperatureTypes
+import com.pradyotprakash.unitconverter.core.models.request.WeightTypes
 import io.ktor.resources.Resource
 
 @Resource(CONVERT)
@@ -14,7 +14,7 @@ class ConverterResource {
     @Resource(LENGTH)
     data class LengthResource(
         private val parent: ConverterResource = ConverterResource(),
-        val value: Long,
+        val value: Double,
         val from: LengthTypes,
         val to: LengthTypes,
     )
@@ -22,7 +22,7 @@ class ConverterResource {
     @Resource(WEIGHT)
     data class WeightResource(
         private val parent: ConverterResource = ConverterResource(),
-        val value: Long,
+        val value: Double,
         val from: WeightTypes,
         val to: WeightTypes,
     )
@@ -30,7 +30,7 @@ class ConverterResource {
     @Resource(TEMPERATURE)
     data class TemperatureResource(
         private val parent: ConverterResource = ConverterResource(),
-        val value: Long,
+        val value: Double,
         val from: TemperatureTypes,
         val to: TemperatureTypes,
     )
